+++ src/pages/admin/comments.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-
 import { useMutation, useQueryClient } from 'react-query';
 
 import { useCommentListQuery } from '../../api/hook/admin';
 import { deleteComment } from '../../api/fetcher/admin';
+import { useIsValidAdmin } from '../../hooks/useIsValidAdmin';
 
 import HomeLayout from '../../components/common/layout/layout';
 import { Container, Title, Section, Tr, Td, PageContent } from '../../styles/admin/user';
@@ -11,11 +11,11 @@ import Button from '../../components/common/button';
 import ArrowLeftIcon from '../../components/common/@Icons/arrowLeftIcon';
 import ArrowRightIcon from '../../components/common/@Icons/arrowRightIcon';
 import GarbageIcon from '../../components/common/@Icons/garbageIcon';
-import AdminRoute from '../../components/common/routes/admin';
 
 const Comments = () => {
   const [page, setPage] = React.useState<number>(1);
   const [totalPage, setTotalPage] = React.useState<number>(0);
+  const isAdmin = useIsValidAdmin();
 
   const queryClient = useQueryClient();
   const { data: commentList } = useCommentListQuery({
@@ -25,6 +25,7 @@ const Comments = () => {
       onSuccess: data => {
         setTotalPage(Math.ceil(data.count / 10));
       },
+      enabled: isAdmin,
     },
   });
 
@@ -76,9 +77,5 @@ const Comments = () => {
 export default Comments;
 
 Comments.getLayout = function getLayout(page: React.ReactElement) {
-  return (
-    <HomeLayout>
-      <AdminRoute>{page} </AdminRoute>
-    </HomeLayout>
-  );
+  return <HomeLayout>{page}</HomeLayout>;
 };
